Added a note
When installing the Android SDK Tools, the versions referred to in the README might not be available if the SDK Manager needs to be updated. In this case, you can use the older SDK Manager to install the newer SDK Manager by just installing the highest version of the Android SDK Tools that you can.